+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Round percents in dashboard
 * Add custom reservation scope for campus
 * Add red dot to notify created rides regulator
+* Luggage visibility increase in planificator
 * Added toast messages on phones & cars edit/creation
 * Remove logs menu, using Loki stack right now
 * Export dashboard and monitoring on deploy